+++ src/pages/Search/Search.tsx
@@ -1,27 +1,12 @@
 /* eslint-disable no-console */
-import React from 'react'
-import { useTheme } from '@material-ui/core/styles'
-import Drawer from '@material-ui/core/Drawer'
-import List from '@material-ui/core/List'
 import Typography from '@material-ui/core/Typography'
-import Divider from '@material-ui/core/Divider'
-import IconButton from '@material-ui/core/IconButton'
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft'
-import ChevronRightIcon from '@material-ui/icons/ChevronRight'
-import ListItem from '@material-ui/core/ListItem'
-import ListItemIcon from '@material-ui/core/ListItemIcon'
-import ListItemText from '@material-ui/core/ListItemText'
-import SearchIcon from '@material-ui/icons/Search'
-import HistoryIcon from '@material-ui/icons/History'
-import BookmarksIcon from '@material-ui/icons/Bookmarks'
-import CommentIcon from '@material-ui/icons/Comment'
-import SettingsIcon from '@material-ui/icons/Settings'
+import React from 'react'
 import styled from 'styled-components'
 import AppBar from '../../components/Layout/AppBar'
+import ExpandableDrawer from '../../components/Layout/ExpandableDrawer'
 
 export default function MiniDrawer(): JSX.Element {
-  const theme = useTheme()
-  const [open, setOpen] = React.useState(false)
+  const [open, setOpen] = React.useState(true)
 
   const handleDrawerOpen = () => {
     console.log(open)
@@ -36,53 +21,7 @@ export default function MiniDrawer(): JSX.Element {
   return (
     <>
       <AppBar open={open} handleOpen={handleDrawerOpen} />
-      <SideDrawer open={open} variant="permanent">
-        <CloseMenuIcon>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === 'rtl' ? (
-              <ChevronRightIcon />
-            ) : (
-              <ChevronLeftIcon />
-            )}
-          </IconButton>
-        </CloseMenuIcon>
-        <Divider />
-        <List>
-          <ListItem button>
-            <ListItemIcon>
-              <SearchIcon />
-            </ListItemIcon>
-            <ListItemText primary={'検索'} />
-          </ListItem>
-          <ListItem button>
-            <ListItemIcon>
-              <BookmarksIcon />
-            </ListItemIcon>
-            <ListItemText primary={'ブックマーク'} />
-          </ListItem>
-          <ListItem button>
-            <ListItemIcon>
-              <HistoryIcon />
-            </ListItemIcon>
-            <ListItemText primary={'履歴'} />
-          </ListItem>
-          <ListItem button>
-            <ListItemIcon>
-              <CommentIcon />
-            </ListItemIcon>
-            <ListItemText primary={'コメント'} />
-          </ListItem>
-        </List>
-        <Divider />
-        <List>
-          <ListItem button>
-            <ListItemIcon>
-              <SettingsIcon />
-            </ListItemIcon>
-            <ListItemText primary={'設定'} />
-          </ListItem>
-        </List>
-      </SideDrawer>
+      <ExpandableDrawer open={open} handleClose={handleDrawerClose} />
       <MainContent>
         <SpacerDiv />
         <Typography paragraph>
@@ -119,64 +58,11 @@ export default function MiniDrawer(): JSX.Element {
   )
 }
 
-const drawerWidth = 240
-
 const SpacerDiv = styled.div`
   offset: ${(props) => props.theme.mixins.toolbar};
-`
-
-const SideDrawer = styled(({ ...rest }) => (
-  // classes paper を上書きする
-  <Drawer variant="permanent" {...rest} classes={{ paper: 'drawer-paper' }} />
-))`
-  flex-shrink: 0;
-  white-space: nowrap;
-  width: ${(props) =>
-    props.open ? drawerWidth : props.theme.spacing(7) + 1}px;
-  transition: ${(props) =>
-    props.open
-      ? props.theme.transitions.create('width', {
-          easing: props.theme.transitions.easing.sharp,
-          duration: props.theme.transitions.duration.enteringScreen
-        })
-      : props.theme.transitions.create('width', {
-          easing: props.theme.transitions.easing.sharp,
-          duration: props.theme.transitions.duration.leavingScreen
-        })};
-  overflow-x: ${(props) => !props.open && 'hidden'};
-  ${(props) => props.theme.breakpoints.up('sm')} {
-    width: theme.spacing(9) + 1;
-  }
-
-  & .drawer-paper {
-    width: ${(props) =>
-      props.open ? drawerWidth : props.theme.spacing(7) + 1}px;
-    transition: ${(props) =>
-      props.open
-        ? props.theme.transitions.create('width', {
-            easing: props.theme.transitions.easing.sharp,
-            duration: props.theme.transitions.duration.enteringScreen
-          })
-        : props.theme.transitions.create('width', {
-            easing: props.theme.transitions.easing.sharp,
-            duration: props.theme.transitions.duration.leavingScreen
-          })};
-    overflow-x: 'hidden';
-    ${(props) => props.theme.breakpoints.up('sm')} {
-      width: theme.spacing(9) + 1;
-    }
-  }
 `
 
 const MainContent = styled.main`
   flex-grow: 1;
   padding: ${(props) => props.theme.spacing(3)}px;
-`
-
-const CloseMenuIcon = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  padding: ${(props) => props.theme.spacing(0, 1)};
-  offset: ${(props) => props.theme.mixins.toolbar};
 `
